+++ src/components/ui/signupForm.tsx
@@ -9,7 +9,7 @@ import { Input } from "@/components/ui/input";
 import { Label } from "@/components/ui/label";
 import { Checkbox } from "@/components/ui/checkbox";
 import { Button } from "@/components/ui/button";
-import PasswordInput from "../../components/ui/passwordInput";
+import PasswordInput from "./passwordInput";
 
 import { useState } from "react";
 import { useRouter } from "next/navigation";
